Move all non-Triss classes to packages
