+++ src/main/java/triss/command/ByeCommand.java
@@ -1,9 +1,9 @@
 package triss.command;
 
-import triss.Parser;
-import triss.Storage;
-import triss.TaskList;
-import triss.Ui;
+import triss.parser.Parser;
+import triss.storage.Storage;
+import triss.tasklist.TaskList;
+import triss.ui.Ui;
 
 public class ByeCommand extends Command {
