+++ src/main/java/triss/tasklist/TaskList.java
@@ -1,6 +1,8 @@
-package triss;
+package triss.tasklist;
 
+import triss.ui.Ui;
 import triss.exception.TrissException;
+import triss.parser.Parser;
 import triss.task.Deadline;
 import triss.task.Event;
 import triss.task.Task;
