+++ src/main/java/triss/ui/Ui.java
@@ -1,4 +1,4 @@
-package triss;
+package triss.ui;
 
 import java.util.Scanner;
